Directory searching and verbosity
It can now accept not only a single document, but also a directory to search for documents in. This allows for cleaning many documents at once without having to run the command for each document individually.
Additionally, I added a verbosity option, which changes how much information is printed when running the script.