Trash Disposal & Improved Error Checking
Added disposal argument which takes the place of the overwrite argument and allows the option to move original documents to trash.
Added code blocks to check for various errors in the arguments and print accurate error messages with tips for solving them when applicable. This will hopefully make it easier to use the script and help clear things up when something doesn't work.